moved LanguagePickerContainer to its own component
